Change Android mv() back to using the system renameTo() method.
A previous change to rn-fetch-blob's mv implementation entailed copying the source file to the destination file and then deleting the source file.  This approach has an undesirable impact on the mv operation, where the operation takes orders of magnitude longer to complete.  It also adds disk space requirements that shouldn't be necessary, since both files need to exist at the same time.
This change goes back to using the renameTo() method, while adding some extra error checking to hopefully alleviate whatever problems caused the original fix.
Note that the POSIX mv() and rename() methods both stipulate that moving to a non-existent directory should be treated as an error.